visible complete button to the assignee user
